Fix impact report
In determining the impact of the latest transaction, the process is to simply compare projections with, and without, the latest transaction included.

However, this could lead to a case where the latest transaction would be inferred as being an entirely new position (adding a bunch to annual income). This happens because when excluding it, projections might show that the position has been closed (depending on date, schedule etc).