Don't implicitly consider positions closed
The problem with implicitly considering a position closed once an expected date for a forecasted transaction is passed, is that it just disappears. There is no obvious indications as to why your forecasted income just dropped.

By having the transactions stick around (and explicitly highlighting them through color), we make it obvious and force you to deal with it. Did you actually close the position, or did the company eliminate their dividend?

A transaction will stick around for 12 months, as that is the cut-off point for forecasts (see `scheduled_transactions(..)`).