Add support for estimate transactions
Any journal entry that only provides a dividend (e.g. no amount) is considered to be an estimate transaction.

This is useful in cases where you, for example, record transactions based on ex-date rather than pay-date, but receive payouts in a different currency, in which case you know the dividend, but not the actual amount (due to exchange rate).