Discard forecasts dated too closely to realized transactions
This is an effort to reduce the number of false-positive projections, with a preference toward fewer, probable projections rather than more, potentially incorrect projections.

I'm not sure that this method is the best way, and i'm sure a bunch of scenarios is still unaccounted for, but it does seem to work out in practice.