Print payout date for preliminary transactions
This brings back a previously removed feature, but only for those transactions waiting to be finalized.

This helps a common use-case when tracking by ex-date and you're curious when you can expect a payout, without having to do-over the report in --by-payout-date.

However, it's problematic showing this additional date, because it lengthens each row a great deal, so for now, I opted to only display it when not in detailed reporting mode.