Fix forecasting for closed positions
This (additional) check was previously removed, but it is indeed necessary to discard early and avoid an otherwise infinite loop when filling a dividend schedule (e.g. if we're looking to fill 4 slots, but the latest record shows a closed position, then we're just going on and on, eventually hitting year 10000).