Add option to skip adjusting for splits
I'm on the fence about making this an option. Forecasts could be affected negatively by not adjusting for splits, and i'm not sure there's ever a scenario where you want to see the unadjusted records. But maybe i'm wrong?

In any case, it can at least be useful for debug purposes.

#30