Allow encapsulating included paths in quotes
It's not really needed, but some may prefer following this convention for paths containing spaces.

The entire line following "include" is always read, so even if the path contains spaces the path will not be broken up; i.e. adding quotes is superfluous, but it should not be wrong to do so.