Add more aggressive rules for deeming positional records redundant
This increases the number of cases in which a positional record can be deemed redundant. For example, closers are no longer kept by default.

This change is in line with the primary focus of the program (tracking dividends), as positional records are more about  improving forecasting; once past the interim period between transactions, they no longer have any value.